requirements.txt
absl-py==0.10.0
astunparse==1.6.3
backcall==0.2.0
blessings==1.7
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cloudpickle==1.6.0
colour==0.1.5
cycler==0.10.0
decorator==4.4.2
descartes==1.1.0
easydict==1.9
future==0.18.2
gast==0.3.3
google-auth==1.21.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gpustat==0.6.0
GPUtil==1.4.0
graphviz==0.8.4
grpcio==1.31.0
h5py==2.10.0
horovod==0.20.0
idna==2.10
imageio==2.9.0
importlib-metadata==1.7.0
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
joblib==0.17.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lap==0.4.0
Markdown==3.2.2
matplotlib==3.3.3
motmetrics==1.1.3
mpi4py==3.0.3
mxnet-cu101==1.6.0.post0
networkx==2.5
numpy==1.18.5
numpy-quaternion==2020.11.2.17.0.49
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.44
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.8
protobuf==3.13.0
psutil==5.7.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.2
pyntcloud==0.1.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.4.1
Shapely==1.7.1
six==1.15.0
sklearn==0.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.4
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.11.18
torch @ https://download.pytorch.org/whl/cu101/torch-1.6.0%2Bcu101-cp37-cp37m-linux_x86_64.whl
torchvision @ https://download.pytorch.org/whl/cu101/torchvision-0.7.0%2Bcu101-cp37-cp37m-linux_x86_64.whl
tornado==6.0.4
tqdm==4.53.0
traitlets==5.0.5
typing==3.7.4.3
typing-extensions==3.10.0.0
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
